+++ src/TgBotLib/Abstracts/ThumbnailMimeType.php
@@ -1,0 +1,10 @@
+<?php
+
+    namespace TgBotLib\Abstracts;
+
+    abstract class ThumbnailMimeType
+    {
+        const Jpeg = 'image/jpeg';
+        const Gif = 'image/gif';
+        const Mp4 = 'video/mp4';
+    }